fix: 당첨 여부를 prize 클래스로 책임 분리
